整合
整合李灏和铭鸿

+++ src/main/java/com/dgut/cloud_disk/pojo/CdstorageUser.java
@@ -18,7 +18,7 @@ public class CdstorageUser implements Serializable {
      * 工号
      */
     @Column(name = "`USER_WORK_ID`")
-    private Long userWorkId;
+    private Integer userWorkId;
 
     /**
      * 用户名
@@ -117,7 +117,7 @@ public class CdstorageUser implements Serializable {
      *
      * @return USER_WORK_ID - 工号
      */
-    public Long getUserWorkId() {
+    public Integer getUserWorkId() {
         return userWorkId;
     }
 
@@ -126,7 +126,7 @@ public class CdstorageUser implements Serializable {
      *
      * @param userWorkId 工号
      */
-    public void setUserWorkId(Long userWorkId) {
+    public void setUserWorkId(Integer userWorkId) {
         this.userWorkId = userWorkId;
     }
